[SPARK-49699][SS] Disable PruneFilters for streaming workloads
The PR proposes to disable PruneFilters if the predicate of the filter is evaluated to `null` / `false` and the filter (and subtree) is streaming.

PruneFilters replaces the `null` / `false` filter with an empty relation, which means the subtree of the filter is also lost. The optimization does not care about whichever operator is in the subtree, hence some important operators like stateful operator, watermark node, observe node could be lost.

The filter could be evaluated to `null` / `false` selectively among microbatches in various reasons (one simple example is the modification of the query during restart), which means stateful operator might not be available for batch N and be available for batch N + 1. For this case, streaming query will fail as batch N + 1 cannot load the state from batch N, and it's not recoverable in most cases.

See new tests in StreamingQueryOptimizationCorrectnessSuite for details.

No.

UT.

No.

Closes #48149 from n-young-db/n-young-db/disable-streaming-prune-filters.

Lead-authored-by: Nick Young <nick.young@databricks.com>
Co-authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
(cherry picked from commit 46b0210edb4ef8490ee4bbc4a40baf202a531b33)
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>