[SPARK-49752][YARN] Remove workaround for YARN-3350
### What changes were proposed in this pull request?

Remove the logic of forcibly setting the log level to WARN for `org.apache.hadoop.yarn.util.RackResolver`.

### Why are the changes needed?

The removed code was introduced in SPARK-5393 as a workaround for YARN-3350, which is already fixed on the YARN 2.8.0/3.0.0.

### Does this PR introduce _any_ user-facing change?

Yes, previously, the log level of RackResolver is hardcoded as WARN even if the user explicitly sets it to DEBUG.

### How was this patch tested?

Review.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48201 from pan3793/SPARK-49752.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>