[SPARK-49282][CONNECT][SQL] Create a shared SparkSessionBuilder interface
### What changes were proposed in this pull request?
This PR adds a shared SparkSessionBuilder interface.

It also adds a SparkSessionCompanion interface which is mean should be implemented by all SparkSession companions (a.k.a. `object SparkSession`. This is currently the entry point for session building, in the future we will also add the management of active/default sessions.

Finally we add a companion for api.SparkSession. This will bind the implementation that is currently located in `org.apache.spark.sql`. This makes it possible to exclusively work with the interface, instead of selecting an implementation upfront.

### Why are the changes needed?
We are creating a shared Scala SQL interface. Building a session is part of this interface.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests. I have added tests for the implementation binding.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48229 from hvanhovell/SPARK-49282.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>