[SPARK-49814][CONNECT] When Spark Connect Client starts, show the `spark version` of the `connect server`
### What changes were proposed in this pull request?
The pr aims to show the spark version of the connect server when Spark Connect Client starts.

### Why are the changes needed?
With the gradual popularize of Spark Connect module, when the Spark Connect client starts, explicitly displaying the spark version of the `connect server`, will reduce confusion for users during execution, such as the new version having some features. However, if it connects to an old version and encounters some problems, it will have to manually troubleshoot.
<img width="878" alt="image" src="https://github.com/user-attachments/assets/c93026e8-03e7-4292-bae2-c3a0f4422191">

### Does this PR introduce _any_ user-facing change?
Yes, Connect‘s end-users can intuitively know the `Spark version` on the `server side` when starting the client, reducing confusion.

### How was this patch tested?
Manually check.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48283 from panbingkun/SPARK-49814.

Authored-by: panbingkun <panbingkun@baidu.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>