[SPARK-49845][CORE] Make `appArgs` and `environmentVariables` optional in REST API
### What changes were proposed in this pull request?

This PR aims to make `appArgs` and `environmentVariables` fields optional in REST API.

### Why are the changes needed?

`appArgs` and `environmentVariables` became mandatory due to the Apache Mesos limitation at Spark 2.2.2. Technically, this is a revert of SPARK-22574.
- https://github.com/apache/spark/pull/19966

Since Apache Spark 4.0 removed Mesos support, we don't need these requirements.
- https://github.com/apache/spark/pull/43135

### Does this PR introduce _any_ user-facing change?

No because this is a relaxation of enforcement.

### How was this patch tested?

Pass the CIs with the revised test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48316 from dongjoon-hyun/SPARK-49845.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>