[SPARK-49843][SQL][3.5] Fix change comment on char/varchar columns
### What changes were proposed in this pull request?

Fix the issue in `AlterTableChangeColumnCommand` where changing the comment of a char/varchar column also tries to change the column type to string.

Backporting #48315 to 3.5

### Why are the changes needed?

Because the newColumn will always be a `StringType` even when the metadata says that it was originally char/varchar.

### Does this PR introduce _any_ user-facing change?

Yes, the query will no longer fail when using this code path.

### How was this patch tested?

New query in golden files.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48333 from stefankandic/branch-3.5.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>