Revert "[SPARK-49808][SQL] Fix a deadlock in subquery execution due o lazy vals"
This reverts commit d7abddc454ffef6ac16e8f6df6f601eec621ddfd.

We had offline discussion, and JoshRosen pointed out that:
> The use of LazyTry vs. a non-try Lazy wrappers needs discussion:
LazyTry caches failures. As a result, there is a potential risk of cancellation exceptions being cached here.
This is possibly an issue in case AQE interrupts subquery execution threads during planning, or otherwise can interrupt a subset of a query plan.

Closes #48362 from zhengruifeng/revert_fix_deadlock.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>