[SPARK-49877][SQL] Change classifyException function signature: add isRuntime argument
### What changes were proposed in this pull request?
The proposal is to update the classifyException function so that it can return either ```AnalysisException``` or ```SparkRuntimeException```. This is achieved by adding a new parameter, ```isRuntime```, and modifying the return type to be ```Throwable with SparkThrowable``` for compatibility with both types.

### Why are the changes needed?
The changes are needed to allow the classifyException function to be used in execution part of the code.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Not needed.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48351 from ivanjevtic-db/Change-classify-exception-function-signature.

Lead-authored-by: ivanjevtic-db <ivan.jevtic@databricks.com>
Co-authored-by: Ivan Jevtic <ivan.jevtic@databricks.com>
Co-authored-by: milastdbx <milan.stefanovic@databricks.com>
Co-authored-by: Maxim Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>