Doc: Fix required java versions for spark 3.5
The original description "PySpark requires Java 8 or later" is incorrect since the latest supported version is 17, the downloading page actually correctly states this. Furthermore, as the downloading page states, support for Java 8 is dropped from spark 3.5.0 and later. So I changed the required java versions to either 11 or 17 as indicated on the downloads page