[SPARK-49864][SQL][FOLLOW-UP] Fix default suggestion for binary arithmetic overflow
### What changes were proposed in this pull request?
Improvement on default branch for try suggestion.

### Why are the changes needed?
When we hit default branch in codeGen, we need to return a default value that would specify that we do not know the function, and not just a blank string.

### Does this PR introduce _any_ user-facing change?
Yes.

### How was this patch tested?
No branch hits this behaviour so far, but we need to safeguard from the possible errors.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48450 from mihailom-db/binaryArithmeticOverflowFollowup.

Authored-by: Mihailo Milosevic <mihailo.milosevic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>