[SPARK-49946][CORE] Require an error class in `SparkOutOfMemoryError`
### What changes were proposed in this pull request?
In the PR, I propose to remove the constructors that accept a plan string as an error message, and leave only constructors with the error classes.

### Why are the changes needed?
To migrate the code which uses `SparkOutOfMemoryError` on new error framework.

### Does this PR introduce _any_ user-facing change?
No, it shouldn't because the exception is supposed to raised by Spark.

### How was this patch tested?
By running the modified test suites:
```
$ build/sbt "core/testOnly *ExecutorSuite"
```

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48442 from MaxGekk/req-error-cond-SparkOutOfMemoryError.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>