[SPARK-50029][SQL] Make `StaticInvoke` compatible with the method that return `Any`
### What changes were proposed in this pull request?
The pr aims to make `StaticInvoke` compatible with the method that return `Any`.

### Why are the changes needed?
Currently, our `StaticInvoke` does not support calling the method with a return type signature of `Any`(actually, the type of return value may be `different data type`), while `Invoke` supports it, let's align it.

### Does this PR introduce _any_ user-facing change?
No, only for spark developer.

### How was this patch tested?
- Add new UT.
- Pass GA.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48542 from panbingkun/SPARK-50029.

Authored-by: panbingkun <panbingkun@baidu.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>