[SPARK-49934][SQL] Add implicit cast for accessing collated map with literal
### What changes were proposed in this pull request?

Add a new case in collation type casts for retrieving values from a map where the keys are collated strings. This will allow the literal key to be implicitly cast to the appropriate collation type used in the map.

### Why are the changes needed?

So explicit collate is no longer needed when accessing map elements.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

New unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48423 from stefankandic/fixMapAccess.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>