[SPARK-49902][SQL] Catch underlying runtime errors in `RegExpReplace`
### What changes were proposed in this pull request?

Earlier, runtime errors in underlying libraries were not caught during runtime in the RegExpReplace expression. The underlying errors were thrown directly to the user. For example, it wouldn't be uncommon to see issues like `java.lang.IndexOutOfBoundsException: No group 3`. This PR introduces a change to catch these underlying issues and throw a SparkException instead which details the input on which the exception failed. The new Spark Exception looks something like `org.apache.spark.SparkException: Could not perform regexp_replace for source = <source>, pattern =
<pattern>, replacement = <replacement> and position = <position>`.

### Why are the changes needed?

Two reasons. First, the new exception details which row the given error occurred on, which makes it easier for the user to debug the query or Spark developers to identify bugs. Second, a Spark Exception is generally considered expected behavior indicating that there were no unintended issues in the query's execution.

### Does this PR introduce _any_ user-facing change?

Yes, a better exception is thrown when RegExpReplace fails.

### How was this patch tested?

Unit test in both codegen as well as interpreted mode.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48379 from harshmotw-db/harshmotw-db/regexp_replace_fix.

Lead-authored-by: Harsh Motwani <harsh.motwani@databricks.com>
Co-authored-by: Maxim Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>