Revert "[SPARK-50007][SQL][SS] Provide default values for metrics on observe API when physical node is lost in executed plan"
This reverts commit 0705e102292323551b3f6dfbb7edabc06d31ceb3.

Rationale:
hvanhovell helped to clarify the semantic of the API - it should capture the data itself and not just evaluating against flowing inputs. For example, suppose the query `df.observe("name", count(lit(1)).as("rc")).filter(false)`, the intention of the API is to "rc" to have 4 if there are 4 rows in df, regardless whether the 4 rows will be soon discarded. The commit 0705e102292323551b3f6dfbb7edabc06d31ceb3 is against this.

Furthermore, if we lost the node due to optimization, users wouldn't know whether Spark somehow dropped the node hence metric is not available, with the commit 0705e102292323551b3f6dfbb7edabc06d31ceb3. Before the commit, users may encounter the issue with unavailability of the metric they just registered, but that is at least playing as distinguishing the case of node loss vs no input.

Closes #48590 from HeartSaVioR/SPARK-50046-revert.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>