[SPARK-50077][SQL] Introduce a new pattern object for LogicalRelation to help avoiding default full params pattern
### What changes were proposed in this pull request?

This PR proposes to introduce a new pattern object to extract BaseRelation and CatalogTable from LogicalRelation, and update existing pattern matches with LogicalRelation with the new pattern.

This PR also adds the information of pattern object and encourage devs to use it.

### Why are the changes needed?

I found it very difficult to add a param into LogicalRelation, because a bunch of places use pattern matching with LogicalRelation, and all of them are using default pattern with specifying all params (despite wildcards).

I looked into the patterns and realized most of the patterns just need relation and catalog table inside the LogicalRelation. output and isStreaming are used infrequently. That said, introducing a new pattern object extracting BaseRelation and CatalogTable from LogicalRelation would cover most of the patterns we have for LogicalRelation, and this would help a lot to introduce a new param into LogicalRelation in future.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

N/A (Existing tests would cover it)

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48605 from HeartSaVioR/SPARK-50077.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>