[SPARK-50093][SQL] Collations that use ICU should have the version of the used ICU library
### What changes were proposed in this pull request?
Changing the versions of the collations that use ICU to be that of the currently used ICU library in spark.

### Why are the changes needed?
Changing the version of the ICU library should automatically change the versions of collations that depend on it. Also, even though the UTF8_LCASE doesn't use ICU for comparisons it still relies on ICU's case mappings between upper and lower  case letters and is for that reason included as well.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added  new unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48626 from stefankandic/changeCollationVersion.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>