[SPARK-49637][SQL] Changed error message for INVALID_FRACTION_OF_SECOND
### What changes were proposed in this pull request?

In this PR, the error message for INVALID_FRACTION_OF_SECOND is changed, such that it no longer suggests turning off ANSI flag. Now, the error suggests using try_ variants of functions for making timestamps. This PR will is done in conjunction with https://github.com/apache/spark/pull/48624, where the try_ variants of these functions are implemented.

### Why are the changes needed?

These changes are needed as part of the effort to move to ANSI by default.

### Does this PR introduce _any_ user-facing change?

Yes, the error message is changed.

### How was this patch tested?

Tests not needed.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #48656 from markonik-db/SPARK-49637-INVALID-FRACTION-OF-SECOND.

Authored-by: Marko Nikacevic <marko.nikacevic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>