[SPARK-49562][SQL] Add SQL pipe syntax for aggregation
### What changes were proposed in this pull request?

This PR adds SQL pipe syntax support for aggregation.

For example:

```
SELECT 1 AS x, 2 AS y, 3 AS z
|> AGGREGATE COUNT(*) GROUP BY x AS z, x + y AS z

1	3	1
```

### Why are the changes needed?

The SQL pipe operator syntax will let users compose queries in a more flexible fashion.

### Does this PR introduce _any_ user-facing change?

Yes, see above.

### How was this patch tested?

This PR adds a few unit test cases, but mostly relies on golden file test coverage. I did this to make sure the answers are correct as this feature is implemented and also so we can look at the analyzer output plans to ensure they look right as well.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #48529 from dtenedor/pipe-aggregate.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>