[SPARK-47094][SQL][TEST][FOLLOWUP] SPJ : fix bucket reducer function
### What changes were proposed in this pull request?
SPJ compatible bucket issue has an implementation of reducible function. This patch fixes the implementation and make it same as in apache iceberg one.

### Why are the changes needed?
With this fix, incompatible number of buckets do not return 1 as GCD, hence the buckets do not reduce to 1 when it used in incompatible number of buckets.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
With unit tests

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47126 from himadripal/fix_spj_transform_expression.

Authored-by: himadripal <hpal@apple.com>
Signed-off-by: huaxingao <huaxin.gao11@gmail.com>