[SPARK-50124][SQL] LIMIT/OFFSET should preserve data ordering
### What changes were proposed in this pull request?

This PR fixes a long-standing correctness bug for LIMIT/OFFSET. Spark has 3 execution paths for LIMIT:
1. If LIMIT is the root node, use `CollectLimitExec` to eliminate the shuffle and get the first N records on the driver side.
2. If there is a global sort under LIMIT, use `TakeOrderedAndProjectExec` to avoid a full sort using top-N algorithm
3. Otherwise, use `GlobalLimitExec` to get the first N records via a single-partition shuffle.

The third execution path has a problem: Spark shuffle reader fetches shuffle blocks in random order and can't preserve the data ordering. It's usually OK when the data order doesn't matter, but the second execution path is not always picked because it has a trigger condition: the LIMIT N must be smaller than TOP_K_SORT_FALLBACK_THRESHOLD .

This bug is more likely to happen for OFFSET because it doesn't have the top-K optimization.

An ideal solution is to have an order preserving mode for the shuffle reader, which is non-trivial work. This PR proposes a workaround: add a local sort after the single-partition shuffle to restore the data ordering.

### Why are the changes needed?

correctness fix

### Does this PR introduce _any_ user-facing change?

Yes, now LIMIT/OFFSET with global sort can preserve the data ordering

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #48661 from cloud-fan/limit.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>