[SPARK-49980][CORE][SQL] Fix potential file stream leaks caused by interruption in canceled tasks
### What changes were proposed in this pull request?

Fix potential file stream leaks caused by interruption in canceled tasks.

### Why are the changes needed?

There is a common pattern of:

```
val intermediateReader = ... constructor ...
intermediateReader.initialize(...) <- this may throw an error, i.e. if the task is interrupted.
return WrapperIterator(intermediateReader)
```

If an error is thrown in `initialize`, `intermediateReader` is never closed and the caller has no reference and cannot close it itself, causing the open stream to leak.

This is handled in many other analogous places, but not everywhere. Fixing some of them here.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48483 from davintjong-db/file-stream-leak.

Authored-by: Davin Tjong <davin.tjong@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>