[SPARK-50033][SQL] Add a hint to logical.Aggregate() node
### What changes were proposed in this pull request?

Adds an abstract hint to the case class of Aggregate(). And adds an extension API to allow custom hint resolution rules.

### Why are the changes needed?

Allow users to define and resolve their own aggregate hints, and use the hints to guide the optimizing/planning behavior.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

a new test for custom aggregate hint

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #48523 from agubichev/agg_hint.

Lead-authored-by: Andrey Gubichev <andrey.gubichev@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>