[SPARK-50163][SS] Fix the RocksDB extra acquireLock release due to the completion listener
### What changes were proposed in this pull request?

Adds a new `releaseForThread(opType, threadRef)` method that only releases the RocksDB acquireLock if the provided threadRef is the same as the threadRef in `acquiredThreadInfo`. Also, modify the task completion listener created in `acquire` to use the new method so that it doesn't accidentally release the lock if it's owned by another thread.

Also some small changes to usages of acquire() and release() in `rollback` and `close` to follow this pattern to avoid extraneous or missed releases due to exceptions/interrupts:
```
acquire()
try {
  ...
} finally {
  release()
}
```

### Why are the changes needed?

There currently exists a race condition in RocksDB where the completion listener may release a lock validly held by another thread, resulting in a thread-unsafe condition. We have seen this bug in production.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added unit tests in RocksDB.scala, all other unit tests continue to pass as expected.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Github Co-pilot 1.5.25.10

Closes #48697 from liviazhu-db/liviazhu-db/rocksdb-acquirelock-fix.

Authored-by: Livia Zhu <livia.zhu@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>